+++ CHANGELOG.md
@@ -17,6 +17,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Remove Qualification Level from admin organisation list
 - Remove deprecated fields from Qualification
 - Move Sectors filter above the Nations filter
+- Allow admin users to edit blank Qualifications when no values have been set from a data import
 
 ## [release-004] - 2022-02-21